% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi_function.R
\name{plot_disaster_map_sea}
\alias{plot_disaster_map_sea}
\alias{plot_disaster_map_world}
\title{Create the map of natural disaster count in South-East Asian countries}
\usage{
plot_disaster_map_sea(disaster_type, year)

plot_disaster_map_world(disaster_type, year)
}
\arguments{
\item{disaster_type}{the type of the disaster}

\item{year}{the year of the disaster}
}
\value{
ggplot A map displaying the count of disasters for each country

ggplot A map displaying the count of disasters for each country
}
\description{
Create the map of natural disaster count in South-East Asian countries

Create the map of natural disaster count for the world
}
\details{
This function focuses specifically on Southeast Asian countries. It uses the
\code{rnaturalearth} package to fetch country shapes and joins this spatial data with
disaster count data from the EM-DAT dataset. Countries without any disasters
of the specified type and year are filled with a default neutral color.

Unlike \code{plot_disaster_map_sea}, which focuses on Southeast Asian countries,
this function provides a global perspective. It uses global country shapes
and disaster data to create a world map. This function is suitable for
broader analyses involving all countries.
}
\author{
ChiMai24
}
