% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mia_function.R
\name{disaster_month}
\alias{disaster_month}
\title{This function analyzes the month with the most frequent peak disaster starts
over a span of multiple years, and produces a bar plot showing the
frequency of peak months}
\usage{
disaster_month(data, location, country)
}
\arguments{
\item{data}{The data from EMDAT that users can freely scale when download}

\item{location}{Name of the place, maybe a city or a district to filter by}

\item{country}{Name of the country the place resides in to filter by}
}
\value{
A statement identifying the most frequent peak months of disaster
starts and a bar graph showing the frequency of peak disaster months.
#' @examples
file_path <- file.path("/Users/tranmaiphuong/Documents/SMITH/SOPHOMORE/SDS 270/public_emdat_2024-11-11.xlsx")
data <- readxl::read_excel(file_path)
disaster_month(data, "Hanoi", "Viet Nam")
}
\description{
This function analyzes the month with the most frequent peak disaster starts
over a span of multiple years, and produces a bar plot showing the
frequency of peak months
}
\author{
Mia Tran
}
