% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdat_world.R
\docType{data}
\name{emdat_world}
\alias{emdat_world}
\title{Emdat World Dataset}
\format{
A data frame with 46 columns:
\describe{
\item{\code{dis_no}}{Unique disaster number.}
\item{\code{historic}}{Indicates if the disaster is classified as historic (logical).}
\item{\code{classification_key}}{Key for disaster classification.}
\item{\code{disaster_group}}{Broad group of disasters (e.g., Natural, Technological).}
\item{\code{disaster_subgroup}}{Subgroup under the disaster group.}
\item{\code{disaster_type}}{Type of disaster (e.g., Flood, Earthquake).}
\item{\code{disaster_subtype}}{Subtype of the disaster.}
\item{\code{external_i_ds}}{External IDs for the disaster.}
\item{\code{event_name}}{Name of the disaster event.}
\item{\code{iso}}{ISO code of the country affected.}
\item{\code{country}}{Country where the disaster occurred.}
\item{\code{subregion}}{Subregion affected.}
\item{\code{region}}{Region affected.}
\item{\code{location}}{Location details.}
\item{\code{origin}}{Origin of the disaster.}
\item{\code{associated_types}}{Associated disaster types.}
\item{\code{ofda_bha_response}}{Indicates if there was a response from OFDA/BHA.}
\item{\code{appeal}}{Indicates if an appeal was issued (logical).}
\item{\code{declaration}}{Declaration related to the disaster.}
\item{\code{aid_contribution_000_us}}{Aid contribution in thousands of USD.}
\item{\code{magnitude}}{Magnitude of the disaster.}
\item{\code{magnitude_scale}}{Scale used for measuring the magnitude.}
\item{\code{latitude}}{Latitude of the disaster location.}
\item{\code{longitude}}{Longitude of the disaster location.}
\item{\code{river_basin}}{River basin affected.}
\item{\code{start_year}}{Year the disaster began.}
\item{\code{start_month}}{Month the disaster began.}
\item{\code{start_day}}{Day the disaster began.}
\item{\code{end_year}}{Year the disaster ended.}
\item{\code{end_month}}{Month the disaster ended.}
\item{\code{end_day}}{Day the disaster ended.}
\item{\code{total_deaths}}{Total deaths caused by the disaster.}
\item{\code{no_injured}}{Number of injured people.}
\item{\code{no_affected}}{Number of people affected.}
\item{\code{no_homeless}}{Number of homeless people.}
\item{\code{total_affected}}{Total number of people affected.}
\item{\code{reconstruction_costs_000_us}}{Reconstruction costs in thousands of USD.}
\item{\code{reconstruction_costs_adjusted_000_us}}{Adjusted reconstruction costs in thousands of USD.}
\item{\code{insured_damage_000_us}}{Insured damage in thousands of USD.}
\item{\code{insured_damage_adjusted_000_us}}{Adjusted insured damage in thousands of USD.}
\item{\code{total_damage_000_us}}{Total damage in thousands of USD.}
\item{\code{total_damage_adjusted_000_us}}{Adjusted total damage in thousands of USD.}
\item{\code{cpi}}{Consumer Price Index used for cost adjustments.}
\item{\code{admin_units}}{Administrative units affected by the disaster.}
\item{\code{entry_date}}{Date the data entry was created.}
\item{\code{last_update}}{Date of the last update to the data entry.}
}
}
\source{
\url{https://public.emdat.be/}
}
\usage{
data(emdat_world)
}
\description{
A dataset containing information about global disaster events, including details about the disaster type, affected regions, and impact metrics.
}
\keyword{datasets}
